Cleaner prod/dev store.
